restoring backend

+++ backend/src/main/java/edu/_2_30_online_banking_system_database/backend/payload/requests/TransactionRequest.java
@@ -1,9 +1,7 @@
 package edu._2_30_online_banking_system_database.backend.payload.requests;
 
-import java.math.BigDecimal;
-
 public record TransactionRequest(
-    BigDecimal amount,
+    Double amount,
     Long fromAccountId,
     Long toAccountId,
     String type,
